requirements.txt
pandas==2.2.3
numpy==2.1.3
sqlalchemy==2.0.36
pyodbc==5.2.0
python-dotenv==1.0.1
scikit-learn==1.5.2
mlxtend==0.23.1
lifetimes==0.11.3
matplotlib==3.9.2
seaborn==0.13.2
pyyaml==6.0.2
plotly==5.24.1
streamlit>=1.41.0,<2.0.0
fastapi==0.116.1
uvicorn==0.35.0
